+++ elisp/chronometrist-sexp.el
@@ -14,6 +14,11 @@ Like `pp', it must accept an OBJECT and optionally a
 STREAM (which is the value of `current-buffer')."
   :type 'function)
 
+(define-derived-mode chronometrist-sexp-mode
+  ;; fundamental-mode
+  emacs-lisp-mode
+  "chronometrist-sexp")
+
 (defmacro chronometrist-sexp-in-file (file &rest body)
   "Run BODY in a buffer visiting FILE, restoring point afterwards."
   (declare (indent defun) (debug t))
@@ -21,16 +26,20 @@ STREAM (which is the value of `current-buffer')."
      (save-excursion ,@body)))
 
 (defmacro chronometrist-loop-file (for expr in file &rest loop-clauses)
-  "`cl-loop' LOOP-CLAUSES over s-expressions in FILE.
+  "`cl-loop' LOOP-CLAUSES over s-expressions in FILE, in reverse.
 VAR is bound to each s-expression."
   (declare (indent defun)
+           (debug nil)
            ;; FIXME
-           (debug ("for" form "in" form &rest &or sexp form)))
+           ;; (debug ("for" form "in" form &rest &or sexp form))
+           )
   `(chronometrist-sexp-in-file ,file
      (goto-char (point-max))
      (cl-loop with ,expr
        while (and (not (bobp))
                   (backward-list)
+                  (or (not (bobp))
+                      (not (looking-at-p "^[[:blank:]]*;")))
                   (setq ,expr (ignore-errors (read (current-buffer))))
                   (backward-list))
        ,@loop-clauses)))
@@ -92,6 +101,8 @@ were none."
   "Create `chronometrist-file' if it doesn't already exist."
   (unless (file-exists-p chronometrist-file)
     (with-current-buffer (find-file-noselect chronometrist-file)
+      (goto-char (point-min))
+      (insert ";;; -*- mode: chronometrist-sexp; -*-")
       (write-file chronometrist-file))))
 
 (cl-defun chronometrist-sexp-new (plist)